requirements.txt
python-telegram-bot==20.3
openai
python-docx
python-pptx
pandas
pdfplumber
PyMuPDF
pydub
Pillow
requests
python-dotenv
matplotlib
beautifulsoup4
lxml
numpy
